Fix #5860

+++ branches/main-develop/webEdition/we/include/we_modules/customer/weCustomer.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * webEdition CMS
  *
@@ -21,11 +22,12 @@
  * @package    webEdition_base
  * @license    http://www.gnu.org/copyleft/gpl.html  GPL
  */
+
 /**
  * General Definition of WebEdition Customer
  *
  */
-class weCustomer extends weModelBase {
+class weCustomer extends weModelBase{
 
 	//properties
 	var $ID;
@@ -44,15 +46,15 @@ class weCustomer extends weModelBase {
 	var $LastAccess;
 	var $ModifyDate;
 	var $ModifiedBy;
-	var $protected = array('ID', 'ParentID', 'Icon', 'IsFolder', 'Path', 'Text','ModifiedBy','ModifyDate');
+	var $protected = array('ID', 'ParentID', 'Icon', 'IsFolder', 'Path', 'Text', 'ModifiedBy', 'ModifyDate');
 	var $properties = array('Username', 'Password', 'Forename', 'Surname', 'LoginDenied', 'MemberSince', 'LastLogin', 'LastAccess', 'AutoLoginDenied', 'AutoLogin');
-		var $udates = array('MemberSince', 'LastLogin', 'LastAccess');
+	var $udates = array('MemberSince', 'LastLogin', 'LastAccess');
 
 	/**
 	 * Default Constructor
 	 * Can load or create new Customer depends of parameter
 	 */
-	function __construct($customerID = 0) {
+	function __construct($customerID = 0){
 
 		$this->table = CUSTOMER_TABLE;
 
@@ -62,58 +64,71 @@ class weCustomer extends weModelBase {
 		$this->LastLogin = 0;
 		$this->LastAccess = 0;
 
-		if ($customerID) {
+		if($customerID){
 			$this->ID = $customerID;
 			$this->load($customerID);
 		}
 	}
 
-	function loadPresistents() {
+	function loadPresistents(){
 		$this->persistent_slots = array();
 		$tableInfo = $this->db->metadata($this->table);
-		for ($i = 0; $i < sizeof($tableInfo); $i++) {
+		for($i = 0; $i < sizeof($tableInfo); $i++){
 			$fname = $tableInfo[$i]["name"];
 			$this->persistent_slots[] = $fname;
-			if (!isset($this->$fname))
+			if(!isset($this->$fname))
 				$this->$fname = '';
 		}
 	}
 
-	function save() {
+/**
+ *	Overwrites an existing User, keeps customerfilters => User should be the same person
+ * @param int $id id to overwrite
+ */
+	function overwrite($id){
+		$tmp = new self($id);
+		$tmp->delete(false);
+		unset($tmp);
+		$this->ID = $id;
+		$this->save();
+	}
+
+	function save(){
 		$this->Icon = "customer.gif";
 		$this->IsFolder = 0;
 		$this->Text = $this->Username;
 		$this->Path = "/" . $this->Username;
 
-		if ($this->MemberSince == 0){
+		if($this->MemberSince == 0){
 			$this->MemberSince = time();
 		}
-		$this->ModifyDate=time();
-		$this->ModifiedBy='backend';
+		$this->ModifyDate = time();
+		$this->ModifiedBy = 'backend';
 
-		$hook = new weHook('customer_preSave', '', array('customer'=>$this,'from'=>'management','type'=>($this->ID ? 'existing':'new')));
-		$ret=$hook->executeHook();
+		$hook = new weHook('customer_preSave', '', array('customer' => $this, 'from' => 'management', 'type' => ($this->ID ? 'existing' : 'new')));
+		$ret = $hook->executeHook();
 
 		weModelBase::save();
 	}
 
 	/**
 	 * delete entry from database
+	 * @param recursive bool if true, customerfilter are deleted as well
 	 */
-	function delete() {
-		if (weModelBase::delete()) {
+	function delete($recursive=true){
+		if(weModelBase::delete() && $recursive){
 			weDocumentCustomerFilter::deleteWebUser($this);
 			return true;
 		}
 		return false;
 	}
 
-	function transFieldName($real_name, &$banche) {
-		if (strpos($real_name,g_l('modules_customer','[other]')!==FALSE)) {
+	function transFieldName($real_name, &$banche){
+		if(strpos($real_name, g_l('modules_customer', '[other]') !== FALSE)){
 			return $real_name;
 		}
 		$pre = explode("_", $real_name);
-		if (($pre[0] != $real_name) && (!in_array($pre[0], $this->protected)) && (!in_array($pre[0], $this->properties))) {
+		if(($pre[0] != $real_name) && (!in_array($pre[0], $this->protected)) && (!in_array($pre[0], $this->properties))){
 			$banche = $pre[0];
 			$field = implode("_", array_slice($pre, 1));
 			return $field;
@@ -121,41 +136,41 @@ class weCustomer extends weModelBase {
 		return $real_name;
 	}
 
-	function getBranches(&$banches, &$fixed, &$other, $mysort='') {
+	function getBranches(&$banches, &$fixed, &$other, $mysort=''){
 
 		$fixed["ID"] = $this->ID; // Bug Fix #8413 + #8520
-		if (isset($this->persistent_slots)) {
+		if(isset($this->persistent_slots)){
 			$orderedarray = $this->persistent_slots;
 			$sortarray = ($mysort != '' ? makeArrayFromCSV($mysort) : range(0, count($orderedarray) - 1));
 
-			if (count($sortarray) != count($orderedarray)) {
+			if(count($sortarray) != count($orderedarray)){
 
-				if (count($sortarray) == count($orderedarray) - 1) {
+				if(count($sortarray) == count($orderedarray) - 1){
 					$sortarray[] = max($sortarray) + 1;
-				} else {
+				} else{
 					$sortarray = range(0, count($orderedarray) - 1);
 				}
 			}
 			$orderedarray = array_combine($sortarray, $orderedarray);
 			ksort($orderedarray);
 
-			foreach ($orderedarray as $per) {
-				$var_value = ((!$this->isnew && isset($this->$per)) ? $var_value=$this->$per : null);
+			foreach($orderedarray as $per){
+				$var_value = ((!$this->isnew && isset($this->$per)) ? $var_value = $this->$per : null);
 
 				$field = $this->transFieldName($per, $branche);
 
-				if ($field != $per) {
+				if($field != $per){
 					$banches[$branche][$field] = $var_value;
-				} else if (in_array($per, $this->properties)) {
+				} else if(in_array($per, $this->properties)){
 					$fixed[$per] = $var_value;
-				} else if (!in_array($per, $this->protected)) {
+				} else if(!in_array($per, $this->protected)){
 					$other[$per] = $var_value;
 				}
 			}
 		}
 	}
 
-	function getBranchesNames() {
+	function getBranchesNames(){
 		$branches = array();
 		$common = array();
 		$other = array();
@@ -165,7 +180,7 @@ class weCustomer extends weModelBase {
 		return array_keys($branches);
 	}
 
-	function getFieldsNames($branch, $mysort='') {
+	function getFieldsNames($branch, $mysort=''){
 		$branches = array();
 		$common = array();
 		$other = array();
@@ -174,25 +189,25 @@ class weCustomer extends weModelBase {
 
 		$arr = array();
 
-		if ($branch == '')
-			$branch = g_l('modules_customer','[other]');
+		if($branch == '')
+			$branch = g_l('modules_customer', '[other]');
 
-		if ($branch == g_l('modules_customer','[common]')) {
-			if (is_array($common)) {
+		if($branch == g_l('modules_customer', '[common]')){
+			if(is_array($common)){
 				$arr = $common;
 			}
-		} else if ($branch == g_l('modules_customer','[other]')) {
-			if (is_array($common)) {
+		} else if($branch == g_l('modules_customer', '[other]')){
+			if(is_array($common)){
 				$arr = $other;
 			}
-		} else {
-			if (isset($branches[$branch]) && is_array($branches[$branch]))
+		} else{
+			if(isset($branches[$branch]) && is_array($branches[$branch]))
 				$arr = $branches[$branch];
 		}
 
 		$ret = array();
-		foreach (array_keys($arr) as $b) {
-			if ($branch == g_l('modules_customer','[other]'))
+		foreach(array_keys($arr) as $b){
+			if($branch == g_l('modules_customer', '[other]'))
 				$ret[$b] = $b;
 			else
 				$ret[$branch . "_" . $b] = $b;
@@ -200,66 +215,66 @@ class weCustomer extends weModelBase {
 		return $ret;
 	}
 
-	function getFieldDbProperties($field_name, $buff=array()) {
+	function getFieldDbProperties($field_name, $buff=array()){
 
-		if (empty($buff)) {
+		if(empty($buff)){
 			$buff = $this->getFieldsDbProperties();
 		}
 
-		foreach ($buff as $b)
-			if ($b["Field"] == $field_name)
+		foreach($buff as $b)
+			if($b["Field"] == $field_name)
 				return $b;
 
 		return array();
 	}
 
-	function getFieldsDbProperties() {
+	function getFieldsDbProperties(){
 		$ret = array();
 		$this->db->query("SHOW COLUMNS FROM " . $this->db->escape($this->table));
-		while ($this->db->next_record()) {
+		while($this->db->next_record()) {
 			$ret[$this->db->f("Field")] = $this->db->Record;
 		}
 
 		return $ret;
 	}
 
-	function isInfoDate($field) {
+	function isInfoDate($field){
 		return in_array($field, $this->udates);
 	}
 
-	function isProperty($field) {
+	function isProperty($field){
 		return in_array($field, $this->properties);
 	}
 
-	function isProtected($field) {
+	function isProtected($field){
 		return in_array($field, $this->protected);
 	}
 
-	function clearSessionVars() {
-		if (isset($_SESSION["customer_session"]))
+	function clearSessionVars(){
+		if(isset($_SESSION["customer_session"]))
 			unset($_SESSION["customer_session"]);
 	}
 
-	function customerNameExist($name) {
+	function customerNameExist($name){
 		$db = new DB_WE();
-		return (f('SELECT 1 AS a FROM ' . CUSTOMER_TABLE . ' WHERE Username="' . $db->escape($name) . '"', 'a', $db)=='1');
+		return (f('SELECT 1 AS a FROM ' . CUSTOMER_TABLE . ' WHERE Username="' . $db->escape($name) . '"', 'a', $db) == '1');
 	}
 
-	function fieldExist($field) {
+	function fieldExist($field){
 		return in_array($field, $this->persistent_slots);
 	}
 
-	function getFieldset() {
+	function getFieldset(){
 		$result = array();
 		$fields = $this->getFieldsDbProperties();
-		foreach ($fields as $k => $v) {
-			if (!$this->isProtected($k))
+		foreach($fields as $k => $v){
+			if(!$this->isProtected($k))
 				$result[] = $k;
 		}
 		return $result;
 	}
 
-	function filenameNotValid() {
+	function filenameNotValid(){
 		return preg_match('|[/]|i', $this->Username);
 	}
